refactor: split factors into returns/beta/scores submodules
